Moved Upgrader code to server implementation.
